[style] #19 - 알림  삭제하기 스타일 수정
